backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
aiosmtplib==3.0.1
jinja2==3.1.3
pydantic[email]==2.5.3
